% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.ar1.R
\name{r.ar1}
\alias{r.ar1}
\title{The auto regressive lag 1 model}
\usage{
r.ar1(n, mu, sigma, phi)
}
\arguments{
\item{n}{The sample length}

\item{mu}{The mean}

\item{sigma}{The Standard Deviation}

\item{phi}{The Correlation coefficient}
}
\value{
A sample of random variables
}
\description{
The auto regressive lag 1 model
}
\examples{
r.ar1(n = 10, mu = 0, sigma = 1, phi = 0.1)

}
